# Improve geometry calculation performance