--- a/docs/src/geolocation.md
+++ b/docs/src/geolocation.md
@@ -4,7 +4,6 @@
 distance
 heading
 calculate_location
-LightOSM.to_matrix
 LightOSM.to_cartesian
 LightOSM.bounding_box_from_point
 ```